# api/openapi.yaml
openapi: 3.0.3
info:
  title: My Hike API
  version: 1.0.0
  description: |
    REST API for My Hike. Except for registration, login, health, and this
    document, endpoints require a JWT access token in the Authorization header.
    Administrator endpoints additionally require the `admin` role.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/activities:
    get:
      tags: [Activities]
      summary: List the authenticated user's activities
      operationId: listActivities
      parameters:
        - $ref: "#/components/parameters/DateQuery"
        - $ref: "#/components/parameters/MonthQuery"
      responses:
        "200":
          description: Activities ordered newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Activities]
      summary: Create an activity
      operationId: createActivity
      requestBody:
        $ref: "#/components/requestBodies/ActivityInput"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/activities/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Activities]
      summary: Get an activity owned by the user; administrators may read any activity
      operationId: getActivity
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Activities]
      summary: Update an owned activity; administrators may update any activity
      operationId: updateActivity
      requestBody:
        $ref: "#/components/requestBodies/ActivityInput"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Activities]
      summary: Delete an owned activity; administrators may delete any activity
      operationId: deleteActivity
      responses:
        "204":
          description: Activity deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/activities/migrate:
    post:
      tags: [Activities]
      summary: Migrate browser activities into the authenticated account
      operationId: migrateActivities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ActivityMigrationItem"
      responses:
        "200":
          description: Migration counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/admin/users:
    get:
      tags: [Administration]
      summary: List users with backend pagination
      description: Administrator role required.
      operationId: listAdminUsers
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - name: search
          in: query
          schema:
            type: string
            maxLength: 320
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/users/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Administration]
      summary: Get a user as administrator
      operationId: getAdminUser
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Administration]
      summary: Update a user's name, email, and role
      operationId: updateAdminUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Administration]
      summary: Soft-delete a user while preserving activities
      operationId: deleteAdminUser
      responses:
        "204":
          description: User deleted; activities preserved
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/users/{id}/activities:
    get:
      tags: [Administration]
      summary: List a user's activities with backend pagination
      operationId: listAdminUserActivities
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: Paginated activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedActivities"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/users/{id}/block:
    put:
      tags: [Administration]
      summary: Block a user
      operationId: blockAdminUser
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Blocked user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/users/{id}/unblock:
    put:
      tags: [Administration]
      summary: Unblock a user
      operationId: unblockAdminUser
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Active user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DateQuery:
      name: date
      in: query
      schema:
        type: string
        format: date
    MonthQuery:
      name: month
      in: query
      schema:
        type: string
        pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        example: "2026-09"
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
  requestBodies:
    ActivityInput:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ActivityInput"
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, expired, or deleted-user token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user is not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found or not owned by the authenticated user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserRole:
      type: string
      enum: [normal_user, admin]
    UserStatus:
      type: string
      enum: [active, blocked, deleted]
    PublicUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/PublicUser"
        - type: object
          required: [status, createdAt]
          properties:
            status:
              $ref: "#/components/schemas/UserStatus"
            createdAt:
              type: string
              format: date-time
    AdminUserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
...
    ActivityType:
      type: string
      enum: [hiking, fitness]
    ActivityInput:
      type: object
      required: [activityType, name, date, minutes, metres, people]
      properties:
        activityType:
          $ref: "#/components/schemas/ActivityType"
        name:
          type: string
          description: Ignored and stored as Fitness when activityType is fitness
        date:
          type: string
          format: date
        minutes:
          type: number
          minimum: 0
        metres:
          type: number
          minimum: 0
          description: Ignored and stored as zero for fitness activities
        people:
          type: array
          minItems: 1
          items:
            type: string
    ActivityMigrationItem:
      allOf:
        - $ref: "#/components/schemas/ActivityInput"
        - type: object
          required: [id]
          properties:
            id:
              type: string
            createdAt:
              type: number
              format: double
    Activity:
      allOf:
        - $ref: "#/components/schemas/ActivityMigrationItem"
        - type: object
          required: [userId, likes, likedBy, slaps, slappedBy]
          properties:
            userId:
              type: string
              format: uuid
            likes:
              type: integer
              minimum: 0
            likedBy:
              type: array
              items:
                type: string
            slaps:
              type: integer
              minimum: 0
            slappedBy:
              type: array
              items:
                type: string
...
    MigrationResult:
      type: object
      required: [imported, skipped]
      properties:
        imported:
          type: integer
          minimum: 0
        skipped:
          type: integer
          minimum: 0
...
    PaginatedUsers:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AdminUser"
    PaginatedActivities:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
    PaginationMetadata:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 1
security:
  - bearerAuth: []